convert participant class into a module
